backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Firebase and database
firebase-admin==6.2.0
google-cloud-firestore==2.13.1

# Data validation and processing
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Environment management
python-dotenv==1.0.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2